Improve MSVC support. Better binary compatibility.
Patch by Patrick Spendrin from KDE.

+++ include/libssh/sftp.h
@@ -56,7 +56,7 @@ extern "C" {
 #endif /* gid_t */
 #ifdef _MSC_VER
 #ifndef ssize_t
-  typedef _W64 signed int ssize_t;
+  typedef _W64 SSIZE_T ssize_t;
 #endif /* ssize_t */
 #endif /* _MSC_VER */
 #endif /* _WIN32 */
